BP5: fix compiler warnings in 32<->64 serializer changes
The CI zero-warning gate failed on the metadata-widening code:
- BP5Serializer.cpp:1348: loop compared size_t against int DimCount
  (-Wsign-compare) -> cast the bound to size_t.
- BP5Serializer.cpp:993: DeferredExtern::BlockID is size_t but
  MetaEntry->BlockCount is now uint64_t, narrowing on 32-bit
  (-Wnarrowing) -> static_cast the value.

Behavior-preserving; DimCount/BlockCount are non-negative counts.

+++ source/adios2/toolkit/format/bp5/BP5Serializer.cpp
@@ -990,7 +990,8 @@ void BP5Serializer::Marshal(void *Variable, const char *Name, const DataType Typ
 
             if (DeferAddToVec)
             {
-                DeferredExterns.push_back({Rec->MetaOffset, MetaEntry->BlockCount - 1, Data,
+                DeferredExterns.push_back({Rec->MetaOffset,
+                                           static_cast<size_t>(MetaEntry->BlockCount - 1), Data,
                                            ElemCount * ElemSize, ElemSize});
             }
             if (Offsets)
@@ -1345,7 +1346,7 @@ void BP5Serializer::CollectFinalShapeValues()
 
             // Shape is fixed-width uint64_t; source Dims is size_t -> element-wise copy
             const auto &VBShape = VB->Shape();
-            for (size_t d = 0; d < Rec->DimCount; d++)
+            for (size_t d = 0; d < static_cast<size_t>(Rec->DimCount); d++)
                 MetaEntry->Shape[d] = static_cast<uint64_t>(VBShape[d]);
         }
     }
